refactor: extracting diagnostic card and status badge from bo_layout
