+++ test/e2e/support/story-urls.json
@@ -17,6 +17,7 @@
   "/storybook/components/core_components_deprecated/table",
   "/storybook/components/data/table",
   "/storybook/components/design_system/breadcrumb",
+  "/storybook/components/design_system/diagnostic_card",
   "/storybook/components/design_system/dropzone",
   "/storybook/components/design_system/ingestion_embedding_banner",
   "/storybook/components/design_system/ingestion_file_browser_header",
@@ -32,10 +33,10 @@
   "/storybook/components/design_system/modal_new_folder",
   "/storybook/components/design_system/modal_rename",
   "/storybook/components/design_system/modal_share",
+  "/storybook/components/design_system/status_badge",
   "/storybook/components/design_system/status_pill",
   "/storybook/components/design_system/toggle",
   "/storybook/components/feedback/flash",
-  "/storybook/components/feedback/status_badge",
   "/storybook/components/file_preview/file_preview",
   "/storybook/components/file_preview/file_preview_modal",
   "/storybook/components/forms/button",
@@ -45,7 +46,6 @@
   "/storybook/components/icons/mcp_endpoint_icons",
   "/storybook/components/misc/channel_capabilities",
   "/storybook/components/misc/config_row",
-  "/storybook/components/misc/diagnostic_card",
   "/storybook/components/misc/header",
   "/storybook/components/pickers/role_share_picker",
   "/storybook/components/pickers/searchable_select",
